requirements.txt
mcp==1.0.0
supabase==2.3.4
python-dotenv==1.0.0
asyncio
typing